Improve the text trigger action
